You can't return from inside a TRY/CATCH/ENDTRY block (see epan/exceptions.h; it crashes).
Try BER first, including both the test dissection and the check of the
results.  If that fails due to an exception being thrown, or due to the
results not indicating that it's a BER-encoded T.125 packet, try PER, so
that if the BER dissection doesn't work we don't just give up.

Get rid of some _U_s attached to parameters that are, in fact, used.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@46066 f5534014-38df-0310-8fa8-9805f1628bb7