We don't need ole32.lib any more; don't link with it and don't include a header file for it.
We also don't monkey around with wide characters any more, so don't
include <tchar.h>.

Depending on the SDK you're using and/or the version of Windows on which
you're building, NDIS_IF_MAX_STRING_SIZE and NETIO_STATUS might not be
defined; define them if not.  Thanks to Gisle Vanem for that fix.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@46150 f5534014-38df-0310-8fa8-9805f1628bb7